Complete 05 exercise.
I thought I had the exercise in the bag, and quickly, but I saw it was
trying to teach me descendant combinator. So I asked on TOP discord to
see if the solution I now have not included in the comments is the
a descendant combinator without a response yet. The first commented
solution was my initial solution. Then I looked at the included
solution and added that below the current code. My first solution did
accomplish the task. And so does the second, disregarding the
difference in browser default text.